requirements.txt
python-dotenv==1.0.0
Faker==19.3.0
pymongo==4.3.3